% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{bad_classification}
\alias{bad_classification}
\title{Detect poor performers}
\usage{
bad_classification(data_, vots = c(0, 70))
}
\arguments{
\item{data_}{Parsed data frame}

\item{vots}{= c(0,70) where to check predicted classification.}
}
\value{
A data.frame of subjects/assignments with column
`exclude80PercentAcc` (whether this assignment is excluded by the criterion
of at least 80% accuracy on 0ms and 70ms VOT). also includes min_correct_logodds.
}
\description{
Basic check: find people who are classifying the endpoints inconsistently or
inaccurately.  One way to measure quality is by the fitted log-odds of
/p/ response at 0 and 70 ms (continuum endpoints, for all intents
and purposes).  This measures how consistent people are (in a
smoothed way) at the ends of the continuum. Intuitively, we want to
throw out anyone who doesn't ever consistently get both /b/ and
/p/.  So we'll set a cutoff of 80% "correct" at either end (meaning
logodds of a /p/ response greater than about 1.4 (`qlogis(0.8)`) at
the /p/ end of 70ms and less than -1.4 at the /b/ end.
}

