% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{bin_trials}
\alias{bin_trials}
\title{Split data_$trial into n bins.}
\usage{
bin_trials(data_, n_bins = 3)
}
\arguments{
\item{data_}{data.frame with at least a `trial` column.}

\item{n_bins}{How many bins to split the trials into (defaults to 3)}
}
\value{
- A data.frame with columns trial (mean trial in bin) and trial_range
(factor with range of trials as strings, ordered correctly).
}
\description{
Split data_$trial into n bins.
}

