% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeling.R
\name{category_boundaries}
\alias{category_boundaries}
\title{Get the category boundary based on an glmer fit and model data}
\usage{
category_boundaries(dat_mod, fit, trials_prop = 5/6)
}
\arguments{
\item{dat_mod}{Model input data (data.frame with at least columns for
bvotCond, bvotCond.s, supervised (numeric supCond predictor), and supCond).}

\item{fit}{Fitted glmer model object}

\item{trials_prop}{(Default: 5/6) which point in the experiment to estimate
the category boundary at.}
}
\value{
data.frame with columns bvotCond, supCond, and boundary_vot
}
\description{
Works by computing predictions at each level of bvotCond and supCond, for
vot predictor of 0 (to get y intercept) and 1 (to get VOT slope).  Then finds
the x intercept, and adjusts for how VOT is scaled and shifted to return the
actual VOT value where the boundary is.
}

