% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excludes}
\alias{excludes}
\title{Excludes}
\format{A data frame with 8 variables:
\describe{
\item{\code{subject}}{Anonymized MTurk worker identifier}
\item{\code{assignmentid}}{Unique ID for the excluded assignment}
\item{\code{lo0ms}}{GLM-extrapolated log-odds of /p/ response at 0ms VOT}
\item{\code{lo70ms}}{GLM-extrapolated log-odds of /p/ response at 70ms VOT}
\item{\code{loMinCorrect}}{The minimum log-odds of a correct response for
  either 0ms or 70ms}
\item{\code{exclude80PercentAcc}}{Logical: whether assignment is excluded
  based on endpoint accuracy}
\item{\code{submittime}}{Time assignment was submitted (from determining
  which assignment was submitted first for subjects with repeats)}
\item{\code{rank}}{Rank of this assignment based on \code{submittime}.
  Greater than 1 means a repeat to be excluded}
}}
\usage{
excludes
}
\description{
Subjects excluded from analysis because they previously participated, or
because they didn't consistently classify the 0ms or 70ms VOT stimuli
sufficiently consistently (as extrapolated by a logistic GLM). Generated
automatically by \code{\link{exclusions}}.
}
\keyword{datasets}

