% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{lhood_to_classification}
\alias{lhood_to_classification}
\title{Convert category likelihoods into categorization function}
\usage{
lhood_to_classification(lhood)
}
\arguments{
\item{lhood}{data frame with likelihoods (see \link{\code{stats_to_lhood}}).}
}
\value{
data frame with columns category and prob_p
}
\description{
Convert category likelihoods into categorization function
}

