% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load_and_parse}
\alias{load_and_parse}
\title{These are functions for pre-processing raw data, including calculating
exclusions based on repeat HITs and poor performance.
Load raw data from .csv file}
\usage{
load_and_parse(filename)
}
\arguments{
\item{filename}{- Filename of the data file to load.}
}
\description{
A couple of steps are requied for reading and cleaning the raw data:
}
\details{
1. Split the block condition information, of the form `visworld_supervised_30`,
   into the useful parts (the overall supervision condition and /b/ VOT)
2. Split the list ID (of the form `BEACH_p_supervised`) into its useful parts,
   which are the word class, actual category, and whether this _trial_ is
   supervised or unsupervised (only matters in the supervised condition)
3. Pull out the category of the image clicked on each trial (and convert to a
   0/1 b/p response variable for later analysis)
4. Convert the /b/ vot condition variable to a factor to thwart ggplot defaults
5. Extract the numeric VOT of the actual stimulus.

The `tidyr::separate` function really comes in handy with the first couple,
and automagically cleans up the data frame by removing the original variabls.
}
\examples{
data <- load_and_parse('data/supunsup-ALL-visworld.csv')
}

