% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeling.R
\name{make_prediction_data}
\alias{make_prediction_data}
\title{Data for prediction}
\usage{
make_prediction_data(data_inc, data_mod)
}
\arguments{
\item{data_inc}{Raw data, what was used to generate the model input}

\item{data_mod}{Model input (raw data at least passed through mutate_for_lmer)}
}
\description{
Generate a data.frame suitable for passing to lme4::predict.merMod.
}

