% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_tools.R
\name{mod_param_init}
\alias{mod_param_init}
\title{create initializer function from data with sensible parameter values}
\usage{
mod_param_init(dat)
}
\description{
create initializer function from data with sensible parameter values
}

