% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{mutate_for_lmer}
\alias{mutate_for_lmer}
\title{Prepare for `glmer`}
\usage{
mutate_for_lmer(.data)
}
\arguments{
\item{.data}{Parsed data frame, with at least columns for vot, trial,
bvotCond, and supCond (e.g., `data(supunsup)`)}
}
\description{
Set center, scale, and set contrasts for lmer modeling.  Variables are
created both for shift-as-continuous predictor (bvotCond.s) and shift-
as-categorical (bvotCond, Helmert Coding).  Likewise for supervision
(supervised and supCond).
}

