% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_tools.R
\name{par_stan}
\alias{par_stan}
\title{Run stan chains in parallel}
\usage{
par_stan(data, file, fit = stan(file, data = data, chains = 0), chains = 4,
  cores = chains, rng_seed = 1, ...)
}
\arguments{
\item{data}{same as `stan`}

\item{file}{filename for a stan model file.  ignored if `fit` also
specified.}

\item{fit}{fitted `stan` object.  Overrides any provided `file` argument.}

\item{chains}{number of chains to run (default: 4).}

\item{cores}{number of cores to use, which controls how many chains can be
run simultaneously. (passed to `mclapply`, so see there for more information)
(default: `chains`)}

\item{rng_seed}{optional, passed to each chain.  (default: 1).}

\item{...}{additional arguments passed to `stan` for each chain.}
}
\value{
A single `stanfit` object with `chains` chains.
}
\description{
Uses `mclapply` as backend, and `sflist2stanfit` to convert to a single,
multi-chain `stanfit` object.  If you don't supply the `fit` parameter
(previously compiled Stan model object) then it will compile one and re-use
it for all the chains.  Returns a single
}

