% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeling.R
\name{plot_category_bounds}
\alias{plot_category_bounds}
\title{Plot category boundaries with error linerange}
\usage{
plot_category_bounds(cat_bounds, dodge_w = 0.75)
}
\arguments{
\item{cat_bounds}{As generated by category_boundaries}

\item{dodge_w}{How much to dodge different conditions by}
}
\description{
Plot category boundaries with error linerange
}

