% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeling.R
\name{predict_and_plot}
\alias{predict_and_plot}
\title{Visualize model fit}
\usage{
predict_and_plot(dat, fit, show_se = FALSE, ...)
}
\arguments{
\item{dat}{Data suitable for calling lme4::predict.merMod (e.g. output of
make_prediction_data).}

\item{fit}{glmer fitted model.}

\item{show_se}{(default: FALSE) whether or not to draw the confidence intervals
(based on the standard error of the fixed effects).}
}
\value{
The ggplot2 plot object.
}
\description{
predict log-odds of /p/ responses based on just the fixed effects of the
fitted model, and plot.
}
\examples{
\dontrun{
data_mod <- mutate_for_lmer(data_inc)
fit <- glmer(respP ~ vot_rel.s * bvotCond.s * supervised * trial.s +
             (vot_rel.s * trial.s | subject),
             data=data_mod, family='binomial',
             control= glmerControl(optimizer='bobyqa'))
data_pred <- make_prediction(data_inc, data_mod)
predict_and_plot(data_pred, fit, show_se=TRUE)
}
}

