% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predict_se}
\alias{predict_se}
\title{Fixed effect prediction SEs}
\usage{
predict_se(fit, dat = NA, mm = mer_model_matrix(fit, dat),
  full_covar = FALSE)
}
\arguments{
\item{fit}{Fitted merMod object}

\item{dat}{Data to generate fixed effects standard errors for}

\item{full_covar}{(Default: FALSE) Whether to return the full covariance matrix
(instead of just the square root of the diagonal elements)}
}
\description{
Generate standard errors for predictions of a mixed effects
model, based on the variance-covariance matrices of the fixed effects.
Returns the standard error, (square root of variance), of the predictions
by default, but can return the full variance-covariance matrix
}

