% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prior_stats}
\alias{prior_stats}
\title{Some reasonable priors on category means/variances}
\format{A data frame with 4 observations of 5 variables:
\describe{
\item{\code{prevoiced}}{Whether or not stats are for prevoiced stops (only for
goldrick et al. data}
\item{\code{mean}}{Mean VOT (in ms) for this category}
\item{\code{sd}}{VOT standard deviation (in ms; inferred or sample)}
\item{\code{category}}{'b' or 'p'}
\item{\code{source}}{Source of this estimate ('goldrick2013' or 'kronrod2012')}
}}
\usage{
prior_stats
}
\description{
These are based on two sources: Kronrod, Copress, and Feldman (2012, CogSci)
used a combination of categorization and discrimination data to estimate
underlying mean and variance for /b/ and /p/, assuming a mean of 60ms for /p/
(which is based on Lisker & Abramson, 1964). Goldrick, Vaughn, and Murphy
(2013, JASA) analyzed VOTs produced by 12 talkers producing 80 monosyllabic
words with word-initial voiced stops (only /b/ analyzed here).  Many talkers
prevoiced some or most of their /b/s, and these are included separately here.
}
\keyword{datasets}

