% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{repeat_subjects}
\alias{repeat_subjects}
\title{Detect repeat subjects}
\usage{
repeat_subjects(data_)
}
\arguments{
\item{data}{- Parsed data frame}
}
\value{
A data.frame with one row per assignment, and the rank of the
submittime of each.  Rank > 1 means it's a repeat for that subject
}
\description{
Detect repeat subjects
}

