% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeling.R
\name{rescale}
\alias{rescale}
\title{Scale, again}
\usage{
rescale(x, x.scaled)
}
\arguments{
\item{x}{vector that needs to be scaled.}

\item{x.caled}{vector that was already scaled.}
}
\description{
Helper function that rescales one variable based on another (previously
scaled), handling presence or absence of centering or scaling appropriately
}

