% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{separatemeans_assignments}
\alias{separatemeans_assignments}
\title{Summary of all Experiment 2 assignments}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124 rows and 10 columns.}
\usage{
separatemeans_assignments
}
\description{
Summary of all Experiment 2 assignments
}
\seealso{
Format described in \code{\link{separatemeans}}
}
\keyword{datasets}

