% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{separatemeans_clean}
\alias{separatemeans_clean}
\title{Data from non-excluded assignments (Expt. 2)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34748 rows and 26 columns.}
\usage{
separatemeans_clean
}
\description{
Data from non-excluded assignments (Expt. 2)
}
\seealso{
Format described in \code{\link{separatemeans}}
}
\keyword{datasets}

