% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{separatemeans_excluded}
\alias{separatemeans_excluded}
\title{Data from excluded assignments (Expt. 2)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1460 rows and 28 columns.}
\usage{
separatemeans_excluded
}
\description{
Data from excluded assignments (Expt. 2)
}
\seealso{
Format described in \code{\link{separatemeans}}
}
\keyword{datasets}

