% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{stats_to_lhood}
\alias{stats_to_lhood}
\title{Convert data frame of category statistics into likelihoods}
\usage{
stats_to_lhood(stats, noise_sd = sd_noise, xlim = c(-30, 90))
}
\arguments{
\item{stats}{data frame with columns category, mean, and sd}

\item{noise_sd}{standard deviation of sensory noise (added, in variance
space, to all category sds)}

\item{xlims}{where to compute liklelihood.}
}
\value{
data frame with columns vot, category, and lhood
}
\description{
Convert data frame of category statistics into likelihoods
}

