% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{str_replace_multi}
\alias{str_replace_multi}
\title{Multiple string substitution}
\usage{
str_replace_multi(str, subsl, replace_all = FALSE)
}
\arguments{
\item{str}{string which needs substitutions}

\item{subsl}{list of substitutions, in the form list(c('orig', 'replace'), ...)}

\item{replace_all}{(default: FALSE) see ?str_replace vs. ?str_replace_all}
}
\description{
Multiple string substitution
}

