% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supunsup_clean}
\alias{supunsup_clean}
\title{Supunsup_clean}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 82362 rows and 27 columns.}
\usage{
supunsup_clean
}
\description{
The clean data set, with excluded assignments removed.
}
\keyword{datasets}

