% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_tools.R
\name{supunsup_to_stan_conj}
\alias{supunsup_to_stan_conj}
\title{convert data to stan input format for the conj_*.stan models}
\usage{
supunsup_to_stan_conj(dat)
}
\description{
convert data to stan input format for the conj_*.stan models
}

