% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_tools.R
\name{supunsup_to_stan_mog}
\alias{supunsup_to_stan_mog}
\title{convert data to stan input format for the mog_*.stan models}
\usage{
supunsup_to_stan_mog(dat)
}
\description{
convert data to stan input format for the mog_*.stan models
}

