% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{bad_classification}
\alias{bad_classification}
\title{Detect poor performers}
\usage{
bad_classification(data_)
}
\arguments{
\item{data_}{- Parsed data frame}
}
\value{
A data.frame of subjects/assignments with columns `lo0ms`, `lo70ms`
(GLM-predicted log odds at 0 and 70ms VOT), `loMinCorrect` (the minimum
log-odds of a correct response, between 0ms and 70ms VOT), and
`exclude80PercentAcc` (whether this assignment is excluded by the criterion
of at least 80% accuracy on 0ms and 70ms VOT).
}
\description{
Basic check: find people who are classifying the endpoints inconsistently or
inaccurately.  One way to measure quality is by the fitted log-odds of
/p/ response at 0 and 70 ms (continuum endpoints, for all intents
and purposes).  This measures how consistent people are (in a
smoothed way) at the ends of the continuum. Intuitively, we want to
throw out anyone who doesn't ever consistently get both /b/ and
/p/.  So we'll set a cutoff of 80% "correct" at either end (meaning
logodds of a /p/ response greater than about 1.4 (`qlogis(0.8)`) at
the /p/ end of 70ms and less than -1.4 at the /b/ end.
}

