% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supunsup_clean}
\alias{supunsup_clean}
\title{Supunsup_clean}
\format{\preformatted{Classes 'tbl_df', 'tbl' and 'data.frame':	82362 obs. of  27 variables:
 $ subject     : Factor w/ 398 levels "supunsup_1","supunsup_10",..: 154 154 154 154 154 154 154 154 154 154 ...
 $ assignmentid: Factor w/ 402 levels "3018Q3ZVOIP7D2ADBHR5X82O49AARI",..: 170 170 170 170 170 170 170 170 170 170 ...
 $ hitid       : Factor w/ 37 levels "31HLTCK4BKUYUG63GZJN68795ZVGVC",..: 11 11 11 11 11 11 11 11 11 11 ...
 $ experiment  : Factor w/ 1 level "supervised-unsupervised-vot": 1 1 1 1 1 1 1 1 1 1 ...
 $ accepttime  : Factor w/ 362 levels "2015-04-02T17:30:52Z",..: 21 21 21 21 21 21 21 21 21 21 ...
 $ submittime  : Factor w/ 396 levels "2015-04-02T17:41:37Z",..: 30 30 30 30 30 30 30 30 30 30 ...
 $ block       : Factor w/ 1 level "visworld": 1 1 1 1 1 1 1 1 1 1 ...
 $ supCond     : Factor w/ 3 levels "mixed","supervised",..: 3 3 3 3 3 3 3 3 3 3 ...
 $ bvotCond    : Factor w/ 5 levels "-10","0","10",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ trial       : int  0 1 2 3 4 5 6 7 8 9 ...
 $ stim        : int  15 26 1 30 26 22 3 31 3 2 ...
 $ stimfn      : Factor w/ 39 levels "BEACH-10.wav",..: 8 28 2 33 28 32 4 34 4 1 ...
 $ wordClass   : Factor w/ 3 levels "BEACH","BEAK",..: 1 3 1 3 3 3 1 3 1 1 ...
 $ respCategory: Factor w/ 2 levels "b","p": 2 1 1 2 1 2 1 2 1 1 ...
 $ trialSupCond: Factor w/ 2 levels "supervised","unsupervised": 2 2 1 2 2 1 1 2 1 1 ...
 $ targetId    : Factor w/ 6 levels "beach","beak",..: 4 3 1 6 3 6 1 6 1 1 ...
 $ targetPos   : Factor w/ 2 levels "left","right": 1 1 2 2 2 2 1 1 1 2 ...
 $ clickx      : int  86 170 485 440 465 432 169 53 126 450 ...
 $ clicky      : int  281 300 328 296 324 299 310 283 309 309 ...
 $ tstart      : num  1.43e+12 1.43e+12 1.43e+12 1.43e+12 1.43e+12 ...
 $ tend        : num  1.43e+12 1.43e+12 1.43e+12 1.43e+12 1.43e+12 ...
 $ rt          : int  11676 952 1446 1367 1285 1576 958 1787 983 1398 ...
 $ respCat     : Factor w/ 2 levels "b","p": 2 1 1 2 1 2 1 2 1 1 ...
 $ respP       : num  1 0 0 1 0 1 0 1 0 0 ...
 $ trueCat     : Factor w/ 2 levels "b","p": 2 1 1 2 1 2 1 2 1 1 ...
 $ vot         : num  40 -20 -20 30 -20 20 0 40 0 -10 ...
 $ labeled     : chr  "unlabeled" "unlabeled" "unlabeled" "unlabeled" ...
}}
\usage{
supunsup_clean
}
\description{
The clean data set, with excluded assignments removed.
}
\keyword{datasets}

