% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supunsup_excluded}
\alias{supunsup_excluded}
\title{Data from excluded assignments (Expt. 1)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6882 rows and 27 columns.}
\usage{
supunsup_excluded
}
\description{
Data from excluded assignments (Expt. 1)
}
\seealso{
Format described in \code{\link{supunsup}}
}
\keyword{datasets}

