% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supunsup_excluded}
\alias{supunsup_excluded}
\title{Supunsup_excluded}
\format{\preformatted{Classes 'tbl_df', 'tbl' and 'data.frame':	6882 obs. of  27 variables:
 $ subject     : Factor w/ 398 levels "supunsup_1","supunsup_10",..: 3 3 3 3 3 3 3 3 3 3 ...
 $ assignmentid: Factor w/ 402 levels "3018Q3ZVOIP7D2ADBHR5X82O49AARI",..: 389 389 389 389 389 389 389 389 389 389 ...
 $ hitid       : Factor w/ 37 levels "31HLTCK4BKUYUG63GZJN68795ZVGVC",..: 6 6 6 6 6 6 6 6 6 6 ...
 $ experiment  : Factor w/ 1 level "supervised-unsupervised-vot": 1 1 1 1 1 1 1 1 1 1 ...
 $ accepttime  : Factor w/ 362 levels "2015-04-02T17:30:52Z",..: 291 291 291 291 291 291 291 291 291 291 ...
 $ submittime  : Factor w/ 396 levels "2015-04-02T17:41:37Z",..: 312 312 312 312 312 312 312 312 312 312 ...
 $ block       : Factor w/ 1 level "visworld": 1 1 1 1 1 1 1 1 1 1 ...
 $ supCond     : Factor w/ 3 levels "mixed","supervised",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ bvotCond    : Factor w/ 5 levels "-10","0","10",..: 3 3 3 3 3 3 3 3 3 3 ...
 $ trial       : int  0 1 2 3 4 5 6 7 8 9 ...
 $ stim        : int  35 14 6 7 27 5 33 43 43 30 ...
 $ stimfn      : Factor w/ 39 levels "BEACH-10.wav",..: 19 9 8 9 31 7 17 18 18 35 ...
 $ wordClass   : Factor w/ 3 levels "BEACH","BEAK",..: 2 1 1 1 3 1 2 2 2 3 ...
 $ respCategory: Factor w/ 2 levels "b","p": 1 2 2 2 1 2 1 1 1 2 ...
 $ trialSupCond: Factor w/ 2 levels "supervised","unsupervised": 1 2 1 1 2 1 1 2 2 2 ...
 $ targetId    : Factor w/ 6 levels "beach","beak",..: 2 4 2 2 3 2 2 5 5 3 ...
 $ targetPos   : Factor w/ 2 levels "left","right": 2 1 2 2 2 1 2 2 2 2 ...
 $ clickx      : int  465 109 462 439 458 164 468 474 463 480 ...
 $ clicky      : int  325 333 320 316 302 313 312 351 334 283 ...
 $ tstart      : num  1.41e+12 1.41e+12 1.41e+12 1.41e+12 1.41e+12 ...
 $ tend        : num  1.41e+12 1.41e+12 1.41e+12 1.41e+12 1.41e+12 ...
 $ rt          : int  2006 5815 1692 1455 1536 1888 1679 1528 1208 1495 ...
 $ respCat     : Factor w/ 2 levels "b","p": 1 2 1 1 1 1 1 2 2 1 ...
 $ respP       : num  0 1 0 0 0 0 0 1 1 0 ...
 $ trueCat     : Factor w/ 2 levels "b","p": 1 2 2 2 1 2 1 1 1 2 ...
 $ vot         : num  20 50 40 50 10 30 0 10 10 50 ...
 $ labeled     : chr  "labeled" "unlabeled" "labeled" "labeled" ...
}}
\usage{
supunsup_excluded
}
\description{
Data from excluded assignments.
}
\keyword{datasets}

